backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-multipart
python-dotenv
pdfplumber
uuid
psycopg2-binary
pydantic
google-generativeai
sse_starlette